Add Simulator/Explorer tab infrastructure
Extract simulator UI into SimulatorTab component. Add ExplorerTab
placeholder with onOpenInSimulator prop for future "Open in Simulator"
flow. App.tsx now renders tab navigation using shadcn/ui Tabs.
Allow underscore-prefixed unused vars in ESLint config.

Closes #43
